+++ ios/sumolog/sumolog/Setting/User/UserSettingViewModel.swift
@@ -44,7 +44,7 @@ class UserSettingViewModel {
         let params = [
             "uuid": uuid,
             "payday": formValues["payday"] as! Int,
-            "price": formValues["price"] as! Int,
+            "price": formValues["price"] as! Double,
             "target_number": formValues["target_number"] as! Int
             ] as [String : Any]
